Edit: Moved shared providers declaration to domain layer
