+++ src/shared/container/providers/HTTPProvider/index.ts
@@ -1,4 +1,4 @@
-import IHTTPProvider from 'shared/container/providers/HTTPProvider/models/IHTTPProvider';
+import IHTTPProvider from 'shared/domain/providers/IHTTPProvider';
 import AxiosHTTPProvider from 'shared/container/providers/HTTPProvider/implementations/AxiosHTTPProvider';
 
 export default function makeHttpProvider(): IHTTPProvider {
