+++ src/shared/container/providers/JwtProvider/implementations/JwtDecodeJwtProvider.ts
@@ -1,5 +1,5 @@
 import decode from 'jwt-decode';
-import { IJwtProvider } from 'shared/container/providers/JwtProvider/models/IJwtProvider';
+import { IJwtProvider } from 'shared/domain/providers/IJwtProvider';
 
 export default class JsonwebtokenJwtProvider implements IJwtProvider {
   public async decode<T = unknown>(token: string): Promise<T | null> {
