+++ src/shared/container/providers/JwtProvider/index.ts
@@ -1,4 +1,4 @@
-import { IJwtProvider } from 'shared/container/providers/JwtProvider/models/IJwtProvider';
+import { IJwtProvider } from 'shared/domain/providers/IJwtProvider';
 import JwtDecodeJwtProvider from './implementations/JwtDecodeJwtProvider';
 
 export default function makeJwtProvider(): IJwtProvider {
