+++ src/shared/container/providers/StorageProvider/index.ts
@@ -1,5 +1,5 @@
 import AsyncStorageProvider from 'shared/container/providers/StorageProvider/implementations/AsyncStorageProvider';
-import IStorageProvider from 'shared/container/providers/StorageProvider/models/IStorageProvider';
+import IStorageProvider from 'shared/domain/providers/IStorageProvider';
 
 export default function makeStorageProvider(): IStorageProvider {
   const storageProvider = new AsyncStorageProvider();
