+++ src/shared/domain/providers/IHTTPProvider.ts
