+++ src/shared/domain/providers/IStorageProvider.ts
